+++ packages/eslint-plugin/lib/configs/rules/stylistic-issues.js
@@ -1,6 +1,34 @@
 // https://eslint.org/docs/rules/#stylistic-issues
 
 module.exports = {
+	// Require or disallow trailing commas
+	// https://eslint.org/docs/rules/comma-dangle
+	'comma-dangle': [ 'error', {
+		arrays: 'always-multiline',
+		objects: 'always-multiline',
+		imports: 'always-multiline',
+		exports: 'always-multiline',
+		functions: 'always-multiline',
+	} ],
+	
+	// Enforce consistent comma style
+	// https://eslint.org/docs/rules/comma-style
+	'comma-style': [ 'error', 'last', {
+		exceptions: {
+			ArrayExpression: false,
+			ArrayPattern: false,
+			ArrowFunctionExpression: false,
+			CallExpression: false,
+			FunctionDeclaration: false,
+			FunctionExpression: false,
+			ImportDeclaration: false,
+			ObjectExpression: false,
+			ObjectPattern: false,
+			VariableDeclaration: false,
+			NewExpression: false,
+		}
+	} ],
+	
 	// Specify whether double or single quotes should be used in JSX attributes
 	// https://eslint.org/docs/rules/jsx-quotes
 	'jsx-quotes': [ 'error', 'prefer-double' ],